+++ CHANGELOG.md
@@ -102,6 +102,7 @@
   - Small fixes to the process of connecting to a WiFi network
   - Refactor SAE Flood Attack - now requires a password to be set as an argument
   - Handle backspace and DEL properly in serial input
+  - Airtag spoofing fixes
   
 - General
   - Disable and re-enable ESP comm manager UART around GPS usage to avoid driver conflicts